% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brownmotion.R
\name{simBM}
\alias{simBM}
\title{Simulate Brownian Motion}
\usage{
simBM(n, sigma = 1)
}
\arguments{
\item{n}{Number of trials}

\item{sigma}{Standard Deviation}
}
\value{
Returns a dataframe with 2 columns "x" and "y" containing the results
of n simulations of Brownian Motion
}
\description{
Input number of trials and sigma and simulate 2D Brownian motion
}
\examples{
simBM(100)
x <- simBM(100)
}

